+++ src/main/java/rocks/zipcode/atm/bank/Bank.java
@@ -14,23 +14,23 @@ public class Bank {
 
     public Bank() {
         accounts.put(1000, new BasicAccount(new AccountData(
-                1000, "Example 1", "example1@gmail.com", 500
+                1000, "Example 1", "example1@gmail.com", 500F
         )));
 
         accounts.put(2000, new PremiumAccount(new AccountData(
-                2000, "Example 2", "example2@gmail.com", 200
+                2000, "Example 2", "example2@gmail.com", 200F
         )));
 
         accounts.put(911, new PremiumAccount(new AccountData(
-                911, "Emergency fund", "pleaseMyFamilyIsStarving@notreal.com", -100
+                911, "Emergency fund", "pleaseMyFamilyIsStarving@notreal.com", -100F
         )));
 
         accounts.put(1, new BasicAccount(new AccountData(
-                1,"John Doe", "ThisDudeIsOld@gmail.com", 10000000
+                1,"John Doe", "ThisDudeIsOld@gmail.com", 10000000F
         )));
 
         accounts.put(2, new PremiumAccount(new AccountData(
-                2,"Theo", "iAmATodler@gmail.com", 0
+                2,"Theo", "iAmATodler@gmail.com", 0F
         )));
     }
 
@@ -44,14 +44,14 @@ public class Bank {
         }
     }
 
-    public ActionResult<AccountData> deposit(AccountData accountData, int amount) {
+    public ActionResult<AccountData> deposit(AccountData accountData, Float amount) {
         Account account = accounts.get(accountData.getId());
         account.deposit(amount);
 
         return ActionResult.success(account.getAccountData());
     }
 
-    public ActionResult<AccountData> withdraw(AccountData accountData, int amount) {
+    public ActionResult<AccountData> withdraw(AccountData accountData, Float amount) {
         Account account = accounts.get(accountData.getId());
         boolean ok = account.withdraw(amount);
